Update grml-2020.06-rc1 release notes
The (list) items of the (highlighted) "New features" were too tiny.
The second level of the unordered list made the fonts so small.

Changed the first level of the unordered list to a (level 4) heading but
did not change the indentation of the (former) second level of the
unordered list to avoid touching too much lines.

While at it also fixed a closing tag of an unordered list.