tips: use tips data from github.com/grml/grml-tips
Removes the old copy from here. Updating tips data needs a redeploy
of the website though; this needs to be remembered.

Since the last update a lot of tips were deleted in grml-tips,
and grml-tips does not keep track of the "tip number". Because of that,
all the numbers change.