Improve handling of line breaks in monospace context
Before, pre was used with its annoying default behavior of overflowing
content instead of breaking at white space line breaking opportunities.

Now, it is defined that white space inside a line is preserved
(remember that HTML elements except <pre> collapse multiple whitespaces
into a since space; so for <pre> we just make it explicit),
but define wrapping behavior when it comes to breaking lines
which is not the default for <pre>.

For the record: these CSS definitions do not destroy single-line
commands when copying text, because visual line breaks are not
represented in copied text.

This can be considered as lightweight fix to the terrible layout
happening on mobile devices with long lines inside <pre> elements.