Get rid of deprecated ways to reference elements in DOM
There used to be a name attribute in HTML which was used to reference
elements in the document structure.

W3C validator says:
  Warning: The name attribute is obsolete. Consider putting an id attribute on the nearest container instead.

The commit …
  • turns name attributes into id attributes
  • gets rid of unused <a> elements

This replacement was done in bulk. Commands like
  %s$><a name="\([^"]\+\)"></a>$ id="\1">$g
have been used.

I did not touch Markdown content though
because Markdown does not provide a mechanism to create IDs for
headings. I am not familiar with the derivative used by hugo which
might introduce some custom syntax to introduce IDs.